requirements.txt
anthropic
pdfplumber
chromadb
fastembed
sentence-transformers
click
python-dotenv
